refactored app.rs into head.rs
